Make tests pass by making stuff eager
It looks like rakudo used to return a List there, but now it returns
a Seq. For some reason (maybe a good one) tests are trying to ensure
that a List is returned, so I added `eager` to satisfy that. That
being said, I think some of these tests should be dropped.